backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth-oauthlib
python-dotenv
cryptography